refactor mcp stdio dto tests
